+++ util/dataFactory/cluster/clusterDataFactory.go
@@ -25,8 +25,8 @@ const (
 	INSERT_QUERY        = "INSERT INTO CLUSTER_MEMBER (namespace, cluster, member_id, member_name, attribute, role, status, createdTime, updatedTime) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)"
 	DELETE_QUERY        = "DELETE FROM CLUSTER_MEMBER WHERE namespace = $1 and cluster = $2 and member_id = $3 and attribute = $4"
 	DELETE_ALL_QUERY    = "DELETE FROM CLUSTER_MEMBER WHERE namespace = $1 and cluster = $2"
-	UPDATE_STATUS_QUERY = "UPDATE CLUSTER_MEMBER SET STATUS = 'invited' WHERE namespace = $1 and cluster = $2 and member_id = $3 and attribute = $4 and updatedTime = $5"
-	UPDATE_ROLE_QUERY   = "UPDATE CLUSTER_MEMBER SET ROLE = '@@ROLE@@' WHERE namespace = $1 and cluster = $2 and member_id = $3 and attribute = $4 and updatedTime = $5"
+	UPDATE_STATUS_QUERY = "UPDATE CLUSTER_MEMBER SET STATUS = 'invited', updatedTime = $1 WHERE namespace = $2 and cluster = $3 and member_id = $4 and attribute = $5 "
+	UPDATE_ROLE_QUERY   = "UPDATE CLUSTER_MEMBER SET ROLE = '@@ROLE@@', updatedTime = $1  WHERE namespace = $2 and cluster = $3 and member_id = $4 and attribute = $5 "
 )
 
 var pg_con_info string
@@ -365,7 +365,7 @@ func ListClusterAllNamespace(userId string, userGroups []string) ([]types.Namesp
 
 	b.WriteString("select namespace, cluster from CLUSTER_MEMBER where 1=1 ")
 
-	b.WriteString("and member_id = '")
+	b.WriteString("and (member_id = '")
 	b.WriteString(userId)
 	b.WriteString("' ")
 
@@ -374,6 +374,7 @@ func ListClusterAllNamespace(userId string, userGroups []string) ([]types.Namesp
 		b.WriteString(userGroup)
 		b.WriteString("' ")
 	}
+	b.WriteString(") ")
 
 	b.WriteString("and status not in ('pending') ")
 
@@ -399,14 +400,14 @@ func ListClusterAllNamespace(userId string, userGroups []string) ([]types.Namesp
 	return clusterManagerNamespacedNameList, nil
 }
 
-func GetPendingUser(clusterMember util.ClusterMemberInfo) ([]util.ClusterMemberInfo, error) {
+func GetPendingUser(clusterMember util.ClusterMemberInfo) (*util.ClusterMemberInfo, error) {
 	db, err := sql.Open("postgres", pg_con_info)
 	if err != nil {
 		klog.Error(err)
 		return nil, err
 	}
 	defer db.Close()
-	clusterMemberList := []util.ClusterMemberInfo{}
+	// clusterMemberList := []util.ClusterMemberInfo{}
 	var b strings.Builder
 
 	b.WriteString("select * from CLUSTER_MEMBER where 1=1 ")
@@ -435,24 +436,22 @@ func GetPendingUser(clusterMember util.ClusterMemberInfo) ([]util.ClusterMemberI
 		return nil, err
 	}
 	defer rows.Close()
-
+	ret := util.ClusterMemberInfo{}
 	if rows.Next() {
-		clusterMember := util.ClusterMemberInfo{}
 		rows.Scan(
-			&clusterMember.Id,
-			&clusterMember.Namespace,
-			&clusterMember.Cluster,
-			&clusterMember.MemberId,
-			&clusterMember.MemberName,
-			&clusterMember.Attribute,
-			&clusterMember.Role,
-			&clusterMember.Status,
-			&clusterMember.CreatedTime,
-			&clusterMember.UpdatedTime,
+			&ret.Id,
+			&ret.Namespace,
+			&ret.Cluster,
+			&ret.MemberId,
+			&ret.MemberName,
+			&ret.Attribute,
+			&ret.Role,
+			&ret.Status,
+			&ret.CreatedTime,
+			&ret.UpdatedTime,
 		)
-		clusterMemberList = append(clusterMemberList, clusterMember)
 	}
-	return clusterMemberList, nil
+	return &ret, nil
 }
 
 func ListPendingUser(cluster string, namespace string) ([]util.ClusterMemberInfo, error) {
@@ -551,7 +550,7 @@ func GetInvitedGroup(clusterMember util.ClusterMemberInfo) (int, error) {
 	return result, nil
 }
 
-func UpdateStatus(item util.ClusterMemberInfo) error {
+func UpdateStatus(item *util.ClusterMemberInfo) error {
 	db, err := sql.Open("postgres", pg_con_info)
 	if err != nil {
 		klog.Error(err)
@@ -562,7 +561,7 @@ func UpdateStatus(item util.ClusterMemberInfo) error {
 	klog.Infoln("Query: " + UPDATE_STATUS_QUERY)
 	klog.Infoln("Paremeters: " + item.Namespace + ", " + item.Cluster + ", " + item.MemberId + ", " + item.Attribute)
 
-	_, err = db.Exec(UPDATE_STATUS_QUERY, item.Namespace, item.Cluster, item.MemberId, item.Attribute, time.Now())
+	_, err = db.Exec(UPDATE_STATUS_QUERY, time.Now(), item.Namespace, item.Cluster, item.MemberId, item.Attribute)
 	if err != nil {
 		klog.Error(err)
 		return err
@@ -583,7 +582,7 @@ func UpdateRole(item util.ClusterMemberInfo) error {
 	klog.Infoln("Query: " + query)
 	klog.Infoln("Paremeters: " + item.Namespace + ", " + item.Cluster + ", " + item.MemberId + ", " + item.Attribute)
 
-	_, err = db.Exec(query, item.Namespace, item.Cluster, item.MemberId, item.Attribute, time.Now())
+	_, err = db.Exec(query, time.Now(), item.Namespace, item.Cluster, item.MemberId, item.Attribute)
 	if err != nil {
 		klog.Error(err)
 		return err
@@ -629,4 +628,52 @@ func DeleteALL(namespace, cluster string) error {
 	}
 
 	return nil
+}
+
+func GetRemainClusterForSubject(namespace, cluster, subject, attribute string) (int, error) {
+	db, err := sql.Open("postgres", pg_con_info)
+	if err != nil {
+		klog.Error(err)
+		return 0, err
+	}
+	defer db.Close()
+	var b strings.Builder
+	var result int
+
+	b.WriteString("select count(*) from  CLUSTER_MEMBER where 1=1 ")
+
+	b.WriteString("and namespace = '")
+	b.WriteString(namespace)
+	b.WriteString("' ")
+
+	b.WriteString("and cluster = '")
+	b.WriteString(cluster)
+	b.WriteString("' ")
+
+	b.WriteString("and member_id = '")
+	b.WriteString(subject)
+	b.WriteString("' ")
+
+	b.WriteString("and attribute = '")
+	b.WriteString(attribute)
+	b.WriteString("' ")
+
+	b.WriteString("and status not in ('pending') ")
+
+	query := b.String()
+	klog.Infoln("Query: " + query)
+	rows, err := db.Query(query)
+
+	if err != nil {
+		klog.Error(err)
+		return 0, err
+	}
+	defer rows.Close()
+
+	if rows.Next() {
+		rows.Scan(
+			&result,
+		)
+	}
+	return result, nil
 }